requirements.txt
fastapi
uvicorn[standard]
python-docx
spacy
ru-core-news-sm==3.7.0
tqdm